+++ include/openrave/openrave.h
@@ -840,7 +840,25 @@ enum IkParameterizationType {
 
     IKP_NumberOfParameterizations=16,     ///< number of parameterizations (does not count IKP_None)
 
-    IKP_UniqueIdMask = 0x000000ff, ///< the mask for the unique ids
+    IKP_VelocityDataBit = 0x00008000, ///< bit is set if the data represents the time-derivate velocity of an IkParameterization
+    IKP_Transform6DVelocity = IKP_Transform6D|IKP_VelocityDataBit,
+    IKP_Rotation3DVelocity = IKP_Rotation3D|IKP_VelocityDataBit,
+    IKP_Translation3DVelocity = IKP_Translation3D|IKP_VelocityDataBit,
+    IKP_Direction3DVelocity = IKP_Direction3D|IKP_VelocityDataBit,
+    IKP_Ray4DVelocity = IKP_Ray4D|IKP_VelocityDataBit,
+    IKP_Lookat3DVelocity = IKP_Lookat3D|IKP_VelocityDataBit,
+    IKP_TranslationDirection5DVelocity = IKP_TranslationDirection5D|IKP_VelocityDataBit,
+    IKP_TranslationXY2DVelocity = IKP_TranslationXY2D|IKP_VelocityDataBit,
+    IKP_TranslationXYOrientation3DVelocity = IKP_TranslationXYOrientation3D|IKP_VelocityDataBit,
+    IKP_TranslationLocalGlobal6DVelocity = IKP_TranslationLocalGlobal6D|IKP_VelocityDataBit,
+    IKP_TranslationXAxisAngle4DVelocity = IKP_TranslationXAxisAngle4D|IKP_VelocityDataBit,
+    IKP_TranslationYAxisAngle4DVelocity = IKP_TranslationYAxisAngle4D|IKP_VelocityDataBit,
+    IKP_TranslationZAxisAngle4DVelocity = IKP_TranslationZAxisAngle4D|IKP_VelocityDataBit,
+    IKP_TranslationXAxisAngleZNorm4DVelocity = IKP_TranslationXAxisAngleZNorm4D|IKP_VelocityDataBit,
+    IKP_TranslationYAxisAngleXNorm4DVelocity = IKP_TranslationYAxisAngleXNorm4D|IKP_VelocityDataBit,
+    IKP_TranslationZAxisAngleYNorm4DVelocity = IKP_TranslationZAxisAngleYNorm4D|IKP_VelocityDataBit,
+
+    IKP_UniqueIdMask = 0x0000ffff, ///< the mask for the unique ids
     IKP_CustomDataBit = 0x00010000, ///< bit is set if the ikparameterization contains custom data, this is only used when serializing the ik parameterizations
 };
 
@@ -1472,10 +1490,11 @@ public:
     /// \brief fills the iterator with the serialized values of the ikparameterization.
     ///
     /// The container the iterator points to needs to have \ref GetNumberOfValues() available.
-    /// Does not support serialized data
+    /// Does not support custom data
+    /// Don't normalize quaternions since it could hold velocity data.
     inline void GetValues(std::vector<dReal>::iterator itvalues) const
     {
-        switch(_type) {
+        switch(_type & ~IKP_VelocityDataBit) {
         case IKP_Transform6D:
             *itvalues++ = _transform.rot.x;
             *itvalues++ = _transform.rot.y;
@@ -1557,11 +1576,11 @@ public:
 
     /// \brief sets a serialized set of values for the IkParameterization
     ///
-    /// Function does not handle custom data
+    /// Function does not handle custom data. Don't normalize quaternions since it could hold velocity data.
     inline void SetValues(std::vector<dReal>::const_iterator itvalues, IkParameterizationType iktype)
     {
         _type = iktype;
-        switch(_type) {
+        switch(_type & ~IKP_VelocityDataBit) {
         case IKP_Transform6D:
             _transform.rot.x = *itvalues++;
             _transform.rot.y = *itvalues++;
@@ -1667,6 +1686,15 @@ public:
         _mapCustomData[name] = values;
     }
 
+    /// \brief sets named custom data in the ik parameterization (\see SetCustomValues)
+    void SetCustomValue(const std::string& name, dReal value)
+    {
+        OPENRAVE_ASSERT_OP_FORMAT0( name.size(), >, 0, "name is empty", ORE_InvalidArguments );
+        OPENRAVE_ASSERT_OP_FORMAT0(std::count_if(name.begin(), name.end(), _IsValidCharInName), ==, (int)name.size(), "name has invalid characters",ORE_InvalidArguments);
+        _mapCustomData[name].resize(1);
+        _mapCustomData[name][0] = value;
+    }
+
     /// \brief gets custom data if it exists, returns false if it doesn't
     bool GetCustomValues(const std::string& name, std::vector<dReal>& values) const
     {
@@ -1701,28 +1729,50 @@ public:
     inline IkParameterization& MultiplyTransform(const Transform& t) {
         switch(GetType()) {
         case IKP_Transform6D:
-            SetTransform6D(t * GetTransform6D());
+            _transform = t * _transform;
+            break;
+        case IKP_Transform6DVelocity:
+            _transform.trans = t.rotate(_transform.trans);
+            _transform.rot = quatMultiply(t.rot,_transform.rot);
             break;
         case IKP_Rotation3D:
-            SetRotation3D(quatMultiply(quatInverse(t.rot),GetRotation3D()));
+        case IKP_Rotation3DVelocity:
+            _transform.rot = quatMultiply(t.rot,_transform.rot);
             break;
         case IKP_Translation3D:
-            SetTranslation3D(t*GetTranslation3D());
+            _transform.trans = t * _transform.trans;
+            break;
+        case IKP_Translation3DVelocity:
+            _transform.trans = t.rotate(_transform.trans);
             break;
         case IKP_Direction3D:
-            SetDirection3D(t.rotate(GetDirection3D()));
+        case IKP_Direction3DVelocity:
+            _transform.rot = t.rotate(_transform.rot);
             break;
         case IKP_Ray4D:
-            SetRay4D(RAY(t*GetRay4D().pos,t.rotate(GetRay4D().dir)));
+            _transform.trans = t * _transform.trans;
+            _transform.rot = t.rotate(_transform.rot);
+            break;
+        case IKP_Ray4DVelocity:
+            _transform.trans = t.rotate(_transform.trans);
+            _transform.rot = t.rotate(_transform.rot);
             break;
         case IKP_Lookat3D:
             SetLookat3D(RAY(t*GetLookat3D(),t.rotate(GetLookat3DDirection())));
             break;
         case IKP_TranslationDirection5D:
-            SetTranslationDirection5D(RAY(t*GetTranslationDirection5D().pos,t.rotate(GetTranslationDirection5D().dir)));
+            _transform.trans = t * _transform.trans;
+            _transform.rot = t.rotate(_transform.rot);
+            break;
+        case IKP_TranslationDirection5DVelocity:
+            _transform.trans = t.rotate(_transform.trans);
+            _transform.rot = t.rotate(_transform.rot);
             break;
         case IKP_TranslationXY2D:
             SetTranslationXY2D(t*GetTranslationXY2D());
+            break;
+        case IKP_TranslationXY2DVelocity:
+            _transform.trans = t.rotate(_transform.trans);
             break;
         case IKP_TranslationXYOrientation3D: {
             Vector v = GetTranslationXYOrientation3D();
@@ -1732,47 +1782,88 @@ public:
             SetTranslationXYOrientation3D(Vector(vnewtrans.y,vnewtrans.y,v.z+zangle));
             break;
         }
+        case IKP_TranslationXYOrientation3DVelocity: {
+            Vector v = GetTranslationXYOrientation3D();
+            Vector voldtrans(v.x,v.y,0);
+            _transform.trans = t.rotate(voldtrans);
+            _transform.trans.z = quatRotate(t.rot,Vector(0,0,v.z)).z;
+            break;
+        }
         case IKP_TranslationLocalGlobal6D:
-            SetTranslationLocalGlobal6D(GetTranslationLocalGlobal6D().first, t*GetTranslationLocalGlobal6D().second);
+            _transform.trans = t*_transform.trans;
+            break;
+        case IKP_TranslationLocalGlobal6DVelocity:
+            _transform.trans = t.rotate(_transform.trans);
             break;
         case IKP_TranslationXAxisAngle4D: {
-            std::pair<Vector,dReal> p = GetTranslationXAxisAngle4D();
-            // don't change the angle since don't know the exact direction it is pointing at
-            SetTranslationXAxisAngle4D(t*p.first,p.second);
+            _transform.trans = t*_transform.trans;
+            // do not support rotations
+            break;
+        }
+        case IKP_TranslationXAxisAngle4DVelocity: {
+            _transform.trans = t.rotate(_transform.trans);
+            // do not support rotations
             break;
         }
         case IKP_TranslationYAxisAngle4D: {
-            std::pair<Vector,dReal> p = GetTranslationYAxisAngle4D();
-            // don't change the angle since don't know the exact direction it is pointing at
-            SetTranslationYAxisAngle4D(t*p.first,p.second);
+            _transform.trans = t*_transform.trans;
+            // do not support rotations
+            break;
+        }
+        case IKP_TranslationYAxisAngle4DVelocity: {
+            _transform.trans = t.rotate(_transform.trans);
+            // do not support rotations
             break;
         }
         case IKP_TranslationZAxisAngle4D: {
-            std::pair<Vector,dReal> p = GetTranslationZAxisAngle4D();
-            // don't change the angle since don't know the exact direction it is pointing at
-            SetTranslationZAxisAngle4D(t*p.first,p.second);
+            _transform.trans = t*_transform.trans;
+            // do not support rotations
             break;
         }
+        case IKP_TranslationZAxisAngle4DVelocity: {
+            _transform.trans = t.rotate(_transform.trans);
+            // do not support rotations
+            break;
+        }
+
         case IKP_TranslationXAxisAngleZNorm4D: {
-            std::pair<Vector,dReal> p = GetTranslationXAxisAngleZNorm4D();
-            // don't change the angle since don't know the exact direction it is pointing at
-            SetTranslationXAxisAngleZNorm4D(t*p.first,p.second);
+            _transform.trans = t*_transform.trans;
+            // only support rotation along z-axis
+            _transform.rot.x -= normalizeAxisRotation(Vector(0,0,1),t.rot).first;
+            break;
+        }
+        case IKP_TranslationXAxisAngleZNorm4DVelocity: {
+            _transform.trans = t.rotate(_transform.trans);
+            // only support rotation along z-axis
+            _transform.rot.x = quatRotate(t.rot,Vector(0,0,_transform.rot.x)).z;
             break;
         }
         case IKP_TranslationYAxisAngleXNorm4D: {
-            std::pair<Vector,dReal> p = GetTranslationYAxisAngleXNorm4D();
-            // don't change the angle since don't know the exact direction it is pointing at
-            SetTranslationYAxisAngleXNorm4D(t*p.first,p.second);
+            _transform.trans = t*_transform.trans;
+            // only support rotation along x-axis
+            _transform.rot.x -= normalizeAxisRotation(Vector(1,0,0),t.rot).first;
+            break;
+        }
+        case IKP_TranslationYAxisAngleXNorm4DVelocity: {
+            _transform.trans = t.rotate(_transform.trans);
+            // only support rotation along x-axis
+            _transform.rot.x = quatRotate(t.rot,Vector(_transform.rot.x,0,0)).x;
             break;
         }
         case IKP_TranslationZAxisAngleYNorm4D: {
-            std::pair<Vector,dReal> p = GetTranslationZAxisAngleYNorm4D();
-            // don't change the angle since don't know the exact direction it is pointing at
-            SetTranslationZAxisAngleYNorm4D(t*p.first,p.second);
+            _transform.trans = t*_transform.trans;
+            // only support rotation along y-axis
+            _transform.rot.x -= normalizeAxisRotation(Vector(0,1,0),t.rot).first;
+            break;
+        }
+        case IKP_TranslationZAxisAngleYNorm4DVelocity: {
+            _transform.trans = t.rotate(_transform.trans);
+            // only support rotation along y-axis
+            _transform.rot.x = quatRotate(t.rot,Vector(0,_transform.rot.x,0)).y;
             break;
         }
         default:
-            throw openrave_exception(str(boost::format("does not support parameterization %d")%GetType()));
+            throw openrave_exception(str(boost::format("does not support parameterization 0x%x")%GetType()));
         }
         for(std::map<std::string, std::vector<dReal> >::iterator it = _mapCustomData.begin(); it != _mapCustomData.end(); ++it) {
             _MultiplyTransform(t, it->first, it->second);
@@ -1842,7 +1933,7 @@ inline IkParameterization operator* (const Transform &t, const IkParameterizatio
         local.SetTransform6D(t * ikparam.GetTransform6D());
         break;
     case IKP_Rotation3D:
-        local.SetRotation3D(quatMultiply(quatInverse(t.rot),ikparam.GetRotation3D()));
+        local.SetRotation3D(quatMultiply(t.rot,ikparam.GetRotation3D()));
         break;
     case IKP_Translation3D:
         local.SetTranslation3D(t*ikparam.GetTranslation3D());
@@ -1910,7 +2001,8 @@ inline IkParameterization operator* (const Transform &t, const IkParameterizatio
         break;
     }
     default:
-        throw openrave_exception(str(boost::format("does not support parameterization %d")%ikparam.GetType()));
+        // internal MultiplyTransform supports more types
+        return IkParameterization(ikparam).MultiplyTransform(t);
     }
     local._mapCustomData = ikparam._mapCustomData;
     for(std::map<std::string, std::vector<dReal> >::iterator it = local._mapCustomData.begin(); it != local._mapCustomData.end(); ++it) {